Moved `CareTypeable` to its own file
